+++ src/main/java/prospectpyxis/pyxislib/utils/ItemUtils.java
@@ -3,10 +3,14 @@ package prospectpyxis.pyxislib.utils;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.Entity;
 import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.RayTraceResult;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
+import prospectpyxis.pyxislib.PyxisLib;
+
+import javax.annotation.Nullable;
 
 public class ItemUtils {
 
@@ -27,6 +31,39 @@ public class ItemUtils {
         double d3 = range;
         Vec3d vec3d1 = vec3d.addVector((double)f6 * d3, (double)f5 * d3, (double)f7 * d3);
         return worldIn.rayTraceBlocks(vec3d, vec3d1, useLiquids, !useLiquids, false);
+    }
+
+    public static boolean compareItems(Item i1, Item against) {
+        if (against == null) return false;
+        else return (i1 == against);
+    }
+
+    public static boolean matchItemStacks(ItemStack i1, ItemStack i2, boolean ignoreMeta) {
+        return (
+                i1.getItem() == i2.getItem() &&
+                (i1.getItemDamage() == i2.getItemDamage() || ignoreMeta)
+        );
+    }
+
+    @Nullable
+    public static ItemStack parseItemStackFromString(String item) {
+        int meta = 0;
+        String[] split = item.split(":");
+        if (split.length < 2 || split.length > 3) {
+            PyxisLib.logger.error("Parsing string " + item + " to ItemStack failed: incorrect amount of colons. Returning null.");
+            return null;
+        }
+        else if (split.length == 3) {
+            meta = Integer.parseInt(split[2]);
+            item = split[0] + ":" + split[1];
+        }
+        Item itemBase = Item.getByNameOrId(item);
+        if (itemBase == null) return null;
+        else return new ItemStack(itemBase, 0, meta);
+    }
+
+    public static boolean itemStackStringHasMeta(String iss) {
+        return iss.split(":").length == 3;
     }
 
     public static class SimpleItem {